+++ examples/benchmarks.py
@@ -62,4 +62,15 @@ chunk(table, 0, 5000)
 print("")
 for doc in table.find('by_multiple', limit=10):
     print('{origin} {day} {hour} {sid} {when}'.format(**doc))
+
+
+start = 0
+count = 5000
+begin = time()
+for doc in table.find('by_multiple'): pass
+finish = time()
+print("  - {:5}:{:5} - Append Speed/sec = {:.0f}".format(start, count, count / (finish - begin)))
+
 db.close()
+
+